backend/requirements.txt
google-adk
amadeus
requests
python-dotenv
flask
flask-cors